// both/models/link.model.ts
import { CollectionObject } from './collection-object.model';

/**
 * interface for Link data model
 *
 * Created by dominiksecka on 3/28/17.
 */
export interface Link extends CollectionObject{
    /**
     * unique identifier generated by MongoDB
     */
    _id: string;
    /**
     * name of link
     */
    name: string;
    /**
     * first direction for the link
     */
    zeroDirection: string;
    /**
     * second direction for the link
     */
    oneDirection: string;
}